11-01-2023

+++ app/src/main/java/com/trc/ccopromo/models/ItemDiscount.java
@@ -1,8 +1,9 @@
 package com.trc.ccopromo.models;
 
+
 public class ItemDiscount{
     public int promoId;
     public String itemCode;
     public String promoName;
-    public int discount;
+    public double discount;
 }
